Adding sections folder
