+++ src/frontend/pages/README.md
@@ -1,0 +1,3 @@
+# Pages Directory
+
+Keeping page files minimal in code and extracting things to sections/<*your page section*> is the way to go.
