+++ src/frontend/sections/README.md
@@ -1,0 +1,3 @@
+# Sections Directory
+
+This directory contains all the sections of the website. A "section" represents some (generally major) feature of the application. All section-specific components can live in here.
